updated RFC-action version to 0.1.0
This includes paritytech/rfc-action#34 which takes into consideration successful RFCs.

When a RFC closed after the last run, it will comment on the open PR (if it is still open) that the RFC has been closed and how to trigger the bot to merge the PR.

If it is succesfull, we can automate it to simply close the PRs without human input.