Sassafras Consensus (#26)
This RFC outlines the core components and procedures of Sassafras
consensus protocol.

---

This RFC does not cover the following topics:

- Implementation details necessary for interoperability: structure of
the on-chain storage, host functions and runtime API.
- Deployment strategies, especially in the context of already existing
blockchain networks
- Parameters required for the zk-SNARK (SRS).
- Networking protocol to deliver tickets to relayers.

These (and probably other) topics are fairly complex and somewhat
independent of the core protocol. They should be the subject of
separate, dedicated RFCs for proper exploration and discussion.

---


[Rendered](https://github.com/davxy/polkadot-fellows-rfcs/blob/sassafras-consensus-rfc/text/0026-sassafras-consensus.md)

Reference implementation tracking issue:
https://github.com/paritytech/polkadot-sdk/issues/41