Introduce a transaction extension version
This RFC proposes to introduce a transaction extensions version. It is proposed to piggyback on [RFC84](https://github.com/paritytech/polkadot-sdk/issues/2415) to not require a new extrinsic format version. With this RFC it will be possible to change the transaction extensions without breaking the extrinsic format of a chain and thus, staying backwards compatible.