Improved light client requests networking protocol (#9)
[Rendered](https://github.com/tomaka/RFCs-1/blob/rfc-9/text/0009-improved-net-light-client-requests.md)

This is a continuation of https://github.com/w3f/PPPs/pull/10

Compared to https://github.com/w3f/PPPs/pull/10, I've changed my mind
about truncated proofs. I think that it's better if the client can fully
verify the proofs that are returned even when the response size limit
would be exceeded. This also means that the proof no longer needs to be
deterministic, and I've removed the paragraph about this.

cc @cheme