XCM Asset Metadata (#125)
This is a refined version of the XCM Asset Metadata RFC.

This PR supersedes the
[PR](https://github.com/polkadot-fellows/xcm-format/pull/50) opened in
the xcm-format repository following the migration of XCM RFCs to the
Fellowship RFCs.

## Summary

This RFC proposes a metadata format for XCM-identifiable assets (i.e.,
for fungible/non-fungible collections and non-fungible tokens) and a set
of instructions to communicate it across chains.

---------

Co-authored-by: Adrian Catangiu <adrian@parity.io>